+++ changelog.md
@@ -6,6 +6,7 @@
 > This version is not released yet and is under active development.
 
 - Fix over-escaping of `[!IMPORTANT]` admonition in changelog.
+- Fix content writing into output files.
 
 ## [4.5.0 (2024-08-24)](https://github.com/kdeldycke/workflows/compare/v4.4.5...v4.5.0)